Consolidate has_admin_scope to access_control.py; remove duplicate from auth.py
has_admin_scope was defined in both api/auth.py and common/access_control.py.
Three API files (github_sources, jobs, telemetry) imported from auth instead
of from the canonical access_control module, creating a layering violation
and divergence risk.

Fix:
- Delete has_admin_scope from auth.py
- Import has_admin_scope from access_control into auth.py for internal use
- Update github_sources.py, jobs.py, telemetry.py to import from access_control
