+++ src/memory/api/telemetry.py
@@ -15,7 +15,8 @@ from sqlalchemy import exists, func
 
 from sqlalchemy.orm import Session as DBSession
 
-from memory.api.auth import get_current_user, has_admin_scope, resolve_user_filter
+from memory.api.auth import get_current_user, resolve_user_filter
+from memory.common.access_control import has_admin_scope
 from memory.common.db.connection import get_session
 from memory.common.db.models import TelemetryEvent, User
 from memory.common.telemetry import (
